+++ corerl/agent/base.py
@@ -101,9 +101,9 @@ class BaseAC(BaseAgent):
         if self.cfg.delta_action:
             self.action_dim = int(self.action_dim / 2)
 
-        self.actor: BaseActor = init_actor(cfg.actor, self.state_dim, self.action_dim)
-        self.q_critic: EnsembleQCritic = init_q_critic(cfg.critic, self.state_dim, self.action_dim)
-        self.v_critic: EnsembleVCritic = init_v_critic(cfg.critic, self.state_dim)
+        self.actor: BaseActor = init_actor(cfg.actor, app_state, self.state_dim, self.action_dim)
+        self.q_critic: EnsembleQCritic = init_q_critic(cfg.critic, app_state, self.state_dim, self.action_dim)
+        self.v_critic: EnsembleVCritic = init_v_critic(cfg.critic, app_state, self.state_dim)
 
     @abstractmethod
     def update_actor(self) -> object: