+++ corerl/component/buffer/base.py
@@ -10,6 +10,7 @@ from torch import Tensor
 from corerl.configs.config import MISSING, config
 from corerl.configs.group import Group
 from corerl.data_pipeline.datatypes import DataMode, StepBatch, Transition, TransitionBatch
+from corerl.state import AppState
 from corerl.utils.device import device
 
 logger = logging.getLogger(__name__)
@@ -21,23 +22,27 @@ class BaseReplayBufferConfig:
     seed: int = MISSING
     memory: int = 1_000_000
     batch_size: int = 256
-    combined: bool = True
-
+    # Whether or not to use combined experience replay:
+    #   https://arxiv.org/pdf/1712.01275
+    # the number of samples in the batch from most recent data.
+    n_most_recent: int = 1
+    id: str = ""
 
 class ReplayBuffer:
-    def __init__(self, cfg: BaseReplayBufferConfig):
+    def __init__(self, cfg: BaseReplayBufferConfig, app_state: AppState):
         self.seed = cfg.seed
         self.rng = np.random.default_rng(self.seed)
         self.memory = cfg.memory
         self.batch_size = cfg.batch_size
+        self.app_state = app_state
 
-        # Whether or not to use combined experience replay:
-        #   https://arxiv.org/pdf/1712.01275
-        self.combined = cfg.combined
+        self.n_most_recent = cfg.n_most_recent
+        assert self.n_most_recent <= self.batch_size
 
         self.data = None
         self.pos = 0
         self.full = False
+        self.id = cfg.id
 
     @abstractmethod
     def _sample_indices(self) -> np.ndarray:
@@ -64,7 +69,16 @@ class ReplayBuffer:
             if not self.full and self.pos == 0:
                 self.full = True
 
+        self.write_buffer_sizes()
         return idxs
+
+    def write_buffer_sizes(self):
+        """
+        Supports writing size of multiple sub buffers
+        """
+        sizes = self.size
+        for i, size in enumerate(sizes):
+            self.app_state.metrics.write(self.app_state.agent_step, metric=f"buffer_{self.id}[{i}]_size", value=size)
 
     def load(self, transitions: Sequence[Transition], data_mode: DataMode) -> np.ndarray:
         assert len(transitions) > 0
@@ -84,19 +98,26 @@ class ReplayBuffer:
 
         return idxs
 
-    def _prepare_sample(self, idxs: np.ndarray) -> list[TransitionBatch]:
+    def get_batch(self, idxs:np.ndarray) -> list[TransitionBatch]:
         if self.size == [0] or self.data is None:
             return []
-
-        if self.combined:
-            idxs[0] = self._last_pos
 
         sampled_data = [self.data[i][idxs] for i in range(len(self.data))]
         return [self._prepare(idxs, sampled_data)]
 
+    def prepare_sample(self, idxs: np.ndarray) -> list[TransitionBatch]:
+        if self.size == [0] or self.data is None:
+            return []
+
+        max_n_most_recent = min(*self.size, self.n_most_recent)
+        for i in range(max_n_most_recent):
+            idxs[i] = self._last_pos-i
+
+        return self.get_batch(idxs)
+
     def sample(self) -> list[TransitionBatch]:
         idxs = self._sample_indices()
-        return self._prepare_sample(idxs)
+        return self.prepare_sample(idxs)
 
     def full_batch(self) -> list[TransitionBatch]:
         if self.size == [0] or self.data is None:
@@ -132,7 +153,7 @@ class ReplayBuffer:
 
 
 buffer_group = Group[
-    [], ReplayBuffer
+    [AppState], ReplayBuffer
 ]()