+++ corerl/component/optimizers/factory.py
@@ -11,6 +11,7 @@ from corerl.component.optimizers.torch_opts import (
     SgdConfig,
     optim_group,
 )
+from corerl.state import AppState
 
 OptimizerConfig = (
     RmspropConfig
@@ -22,7 +23,8 @@ OptimizerConfig = (
 
 def init_optimizer(
     cfg: OptimConfig,
+    app_state: AppState,
     param: Iterable[torch.nn.Parameter],
     ensemble: bool = False,
 ):
-    return optim_group.dispatch(cfg, param, ensemble)
+    return optim_group.dispatch(cfg, app_state, param, ensemble)