+++ corerl/data_pipeline/datatypes.py
@@ -109,6 +109,9 @@ class Transition:
             else:
                 yield attr
 
+    def __len__(self) -> int:
+        return len(self.steps)-1
+
 
 @dataclass
 class StepBatch:
@@ -137,6 +140,18 @@ class StepBatch:
                 and torch.equal(self.dp, other.dp)
         )
 
+    def __getitem__(self, idx: int|slice) -> "StepBatch":
+        if isinstance(idx, int):
+            idx = slice(idx, idx+1)
+
+        return StepBatch(
+            reward=self.reward[idx],
+            action=self.action[idx],
+            gamma=self.gamma[idx],
+            state=self.state[idx],
+            dp=self.dp[idx]
+        )
+
 @dataclass
 class TransitionBatch:
     idxs: np.ndarray
@@ -156,6 +171,17 @@ class TransitionBatch:
                 and torch.equal(self.n_step_gamma, other.n_step_gamma)
         )
 
+    def __getitem__(self, idx: int|slice) -> "TransitionBatch":
+        if isinstance(idx, (int, np.integer)):
+            idx = slice(idx, idx+1)
+
+        return TransitionBatch(
+            idxs=self.idxs[idx],
+            prior=self.prior[idx],
+            post=self.post[idx],
+            n_step_reward=self.n_step_reward[idx],
+            n_step_gamma=self.n_step_gamma[idx]
+        )
 
 class DataMode(Enum):
     OFFLINE = auto()