+++ corerl/interaction/deployment_interaction.py
@@ -4,13 +4,16 @@ from datetime import UTC, datetime, timedelta
 
 import numpy as np
 import pandas as pd
+from torch import Tensor
 
 from corerl.agent.base import BaseAgent
 from corerl.data_pipeline.datatypes import DataMode
 from corerl.data_pipeline.pipeline import Pipeline, PipelineReturn
 from corerl.environment.async_env.async_env import AsyncEnv
 from corerl.environment.async_env.deployment_async_env import DeploymentAsyncEnv
+from corerl.eval.actor_critic import ActorCriticEval
 from corerl.eval.monte_carlo import MonteCarloEvaluator
+from corerl.eval.plotting.evals import plot_evals
 from corerl.interaction.configs import DepInteractionConfig
 from corerl.interaction.interaction import Interaction
 from corerl.messages.events import Event, EventType
@@ -78,7 +81,17 @@ class DeploymentInteraction(Interaction):
             self.restore_checkpoint()
 
         # evals
-        self._monte_carlo_eval = MonteCarloEvaluator(app_state.cfg.eval_cfgs.monte_carlo, app_state, agent)
+        self._monte_carlo_eval = MonteCarloEvaluator(
+            app_state.cfg.eval_cfgs.monte_carlo,
+            app_state,
+            agent,
+        )
+        self._actor_critic_eval = ActorCriticEval(
+            self._app_state.cfg.eval_cfgs.actor_critic,
+            app_state,
+            agent,
+            self._column_desc,
+        )
 
 
     # -----------------------
@@ -91,9 +104,6 @@ class DeploymentInteraction(Interaction):
         pipe_return = self._pipeline(o, data_mode=DataMode.ONLINE)
         self._log_rewards(pipe_return)
         self._agent.update_buffer(pipe_return)
-
-        # perform evaluations
-        self._monte_carlo_eval.execute(pipe_return)
 
         # capture latest state
         state = pipe_return.states
@@ -110,6 +120,18 @@ class DeploymentInteraction(Interaction):
         )
 
         self._write_state_features(state)
+
+        # perform evaluations
+        self._monte_carlo_eval.execute(pipe_return, "online")
+        label = str(self._app_state.agent_step)
+        self._actor_critic_eval.execute([Tensor(self._last_state)], label)
+        plot_evals(
+            app_state=self._app_state,
+            step_start=self._app_state.agent_step,
+            step_end=self._app_state.agent_step,
+            labels=[label]
+        )
+
 
         state_timestamp = state.index[-1]
         if isinstance(state_timestamp, pd.Timestamp):