+++ corerl/interaction/sim_interaction.py
@@ -2,12 +2,15 @@ import logging
 
 import numpy as np
 import pandas as pd
+from torch import Tensor
 
 from corerl.agent.base import BaseAgent
 from corerl.data_pipeline.datatypes import DataMode
 from corerl.data_pipeline.pipeline import Pipeline
 from corerl.environment.async_env.async_env import AsyncEnv
+from corerl.eval.actor_critic import ActorCriticEval
 from corerl.eval.monte_carlo import MonteCarloEvaluator
+from corerl.eval.plotting.evals import plot_evals
 from corerl.interaction.configs import SimInteractionConfig
 from corerl.interaction.interaction import Interaction
 from corerl.messages.events import Event, EventType
@@ -40,7 +43,17 @@ class SimInteraction(Interaction):
         self._last_action: np.ndarray | None = None
 
         # evals
-        self._monte_carlo_eval = MonteCarloEvaluator(app_state.cfg.eval_cfgs.monte_carlo, app_state, agent)
+        self._monte_carlo_eval = MonteCarloEvaluator(
+            app_state.cfg.eval_cfgs.monte_carlo,
+            app_state,
+            agent,
+        )
+        self._actor_critic_eval = ActorCriticEval(
+            self._app_state.cfg.eval_cfgs.actor_critic,
+            app_state,
+            agent,
+            self._column_desc,
+        )
 
 
     # -----------------------
@@ -74,7 +87,15 @@ class SimInteraction(Interaction):
         self._write_to_metrics(pipe_return.rewards)
 
         # perform evaluations
-        self._monte_carlo_eval.execute(pipe_return)
+        self._monte_carlo_eval.execute(pipe_return, "online")
+        label = str(self._app_state.agent_step)
+        self._actor_critic_eval.execute([Tensor(self._last_state)], label)
+        plot_evals(
+            app_state=self._app_state,
+            step_start=self._app_state.agent_step,
+            step_end=self._app_state.agent_step,
+            labels=[label]
+        )
 
         self._app_state.agent_step += 1